stop recording on crash detection
